RCA update test SchoolYear and fix delete of SchoolYear service

+++ src/main/java/com/iesfranciscodelosrios/service/SchoolYearService.java
@@ -28,6 +28,7 @@ public class SchoolYearService implements iServices<SchoolYear> {
 
     @Override
     public SchoolYear delete(SchoolYear schoolYear) {
+        if (schoolYear == null) return null;
         schoolYearRepository.delete(schoolYear);
         return schoolYear;
     }
